fix typo; generate *Details structs
