+++ pkg/sdk/generator/defs/catalog_integrations_def.go
@@ -1,6 +1,7 @@
 package defs
 
 import (
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
 	g "github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk/generator/gen"
 
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk/generator/gen/sdkcommons"
@@ -209,4 +210,52 @@ var catalogIntegrationsDef = g.NewInterface(
 			Describe().
 			SQL("CATALOG INTEGRATION").
 			Name().
-			WithValidation(g.ValidIdentifier, "name"))
+			WithValidation(g.ValidIdentifier, "name"),
+		g.PlainStruct("CatalogIntegrationAwsGlueDetails").
+			AccountObjectIdentifier().
+			Field("CatalogSource", g.KindOfT[sdkcommons.CatalogIntegrationCatalogSourceType]()).
+			Field("TableFormat", g.KindOfT[sdkcommons.CatalogIntegrationTableFormat]()).
+			Bool("Enabled").
+			Number("RefreshIntervalSeconds").
+			Text("Comment").
+			Text("GlueAwsRoleArn").
+			Text("GlueCatalogId").
+			Text("GlueRegion").
+			Text("CatalogNamespace"),
+		g.PlainStruct("CatalogIntegrationObjectStorageDetails").
+			AccountObjectIdentifier().
+			Field("CatalogSource", g.KindOfT[sdkcommons.CatalogIntegrationCatalogSourceType]()).
+			Field("TableFormat", g.KindOfT[sdkcommons.CatalogIntegrationTableFormat]()).
+			Bool("Enabled").
+			Number("RefreshIntervalSeconds").
+			Text("Comment"),
+		g.PlainStruct("CatalogIntegrationOpenCatalogDetails").
+			AccountObjectIdentifier().
+			Field("CatalogSource", g.KindOfT[sdkcommons.CatalogIntegrationCatalogSourceType]()).
+			Field("TableFormat", g.KindOfT[sdkcommons.CatalogIntegrationTableFormat]()).
+			Bool("Enabled").
+			Number("RefreshIntervalSeconds").
+			Text("Comment").
+			Text("CatalogNamespace").
+			Field("RestConfig", g.KindOfT[sdk.OpenCatalogRestConfig]()).
+			Field("RestAuthentication", g.KindOfT[sdk.OAuthRestAuthentication]()),
+		g.PlainStruct("CatalogIntegrationIcebergRestDetails").
+			AccountObjectIdentifier().
+			Field("CatalogSource", g.KindOfT[sdkcommons.CatalogIntegrationCatalogSourceType]()).
+			Field("TableFormat", g.KindOfT[sdkcommons.CatalogIntegrationTableFormat]()).
+			Bool("Enabled").
+			Number("RefreshIntervalSeconds").
+			Text("Comment").
+			Text("CatalogNamespace").
+			Field("RestConfig", g.KindOfT[sdk.IcebergRestRestConfig]()).
+			OptionalField("OAuthRestAuthentication", g.KindOfT[sdk.OAuthRestAuthentication]()).
+			OptionalField("BearerRestAuthentication", g.KindOfT[sdk.BearerRestAuthentication]()).
+			OptionalField("SigV4RestAuthentication", g.KindOfT[sdk.SigV4RestAuthentication]()),
+		g.PlainStruct("CatalogIntegrationSapBdcDetails").
+			AccountObjectIdentifier().
+			Field("CatalogSource", g.KindOfT[sdkcommons.CatalogIntegrationCatalogSourceType]()).
+			Field("TableFormat", g.KindOfT[sdkcommons.CatalogIntegrationTableFormat]()).
+			Bool("Enabled").
+			Number("RefreshIntervalSeconds").
+			Text("Comment"),
+	)
