+++ pkg/sdk/generator/gen/plain_struct.go
@@ -1,5 +1,7 @@
 package gen
 
+import "fmt"
+
 type plainStruct struct {
 	name   string
 	fields []plainField
@@ -23,6 +25,10 @@ func (v *plainStruct) Field(name string, kind string) *plainStruct {
 		kind: kind,
 	})
 	return v
+}
+
+func (v *plainStruct) OptionalField(name string, kind string) *plainStruct {
+	return v.Field(name, fmt.Sprintf("*%s", kind))
 }
 
 func (v *plainStruct) Text(name string) *plainStruct {
